+++ c_src/enacl.h
@@ -9,6 +9,7 @@
 #define ATOM_FALSE "false"
 
 ERL_NIF_TERM enacl_error_tuple(ErlNifEnv *, char *);
+ERL_NIF_TERM enacl_error_finalized(ErlNifEnv *);
 ERL_NIF_TERM enacl_internal_error(ErlNifEnv *);
 
 #endif